Delete mentor endpoint references and replace with user
