+++ src/redux/User.redux/User.reducer.js
@@ -1,13 +1,20 @@
-import { LOG_OUT, SET_USER, GET_USER_INFO } from '../actions';
+import { LOG_OUT, SET_USER, GET_USER_INFO, UPDATE_USER_INFO } from '../actions';
 
 /**
  * User Types are:
  * STUDENT, MENTOR, TEACHER, DIRECTOR, ADVISOR, HEADMASTER
  */
+
 const userInitialState = {
   userType: '',
   firstName: '',
   lastName: '',
+  dateOfBirth: '',
+  timezone: '',
+  phone: '',
+  city: '',
+  languages: '',
+  password: '',
   email: '',
   id: '',
   externalId: '',
@@ -18,6 +25,8 @@ export const user = (state = userInitialState, action) => {
     case SET_USER:
       return action.payload;
     case GET_USER_INFO:
+      return action.payload;
+    case UPDATE_USER_INFO:
       return action.payload;
     case LOG_OUT:
       return userInitialState;
