+++ src/redux/rootReducer.js
@@ -6,7 +6,6 @@ import { authToken } from './AuthToken.redux/AuthToken.reducer';
 import { sessionSlot } from './SessionSlot.redux/SessionSlot.reducer';
 import { registrationForm } from './Registration.redux/RegistrationForm.reducer';
 import { processing } from './Processing.redux/Processing.reducer';
-import { mentor } from './Mentor.redux/Mentor.reducer';
 
 /**
  * Exports all reducers as a single combined reducer
@@ -18,7 +17,6 @@ const rootReducer = combineReducers({
   user,
   registrationForm,
   processing,
-  mentor
 });
 
 export default rootReducer;
